fix(tasks): wait after package upload, not build info upload
