+++ src/main/kotlin/com/liftric/octopusdeploy/task/UploadBuildInformationTask.kt
@@ -1,7 +1,6 @@
 package com.liftric.octopusdeploy.task
 
 import com.liftric.octopusdeploy.shell
-import okhttp3.logging.HttpLoggingInterceptor
 import org.gradle.api.DefaultTask
 import org.gradle.api.provider.Property
 import org.gradle.api.tasks.Input
@@ -38,38 +37,18 @@ open class UploadBuildInformationTask : DefaultTask() {
     @Optional
     var buildInformation: File? = null
 
-    @Input
-    val waitForReleaseDeployments: Property<Boolean> = project.objects.property()
-
-    @Input
-    val waitTimeoutSeconds: Property<Long> = project.objects.property()
-
-    @Input
-    val delayBetweenChecksSeconds: Property<Long> = project.objects.property()
-
-    /**
-     * Configures the http logging of the underlying okhttp client used for octopus api requests
-     */
-    @Input
-    val httpLogLevel: Property<HttpLoggingInterceptor.Level> = project.objects.property()
-
     @TaskAction
     fun execute() {
-        val awaitReleases = waitForReleaseDeployments.getOrElse(false)
-        val packageNameValue = packageName.get()
-        val versionValue = version.get()
-        val octopusUrlValue = octopusUrl.get()
-        val apiKeyValue = apiKey.get()
         val (exitCode, inputText, errorText) = listOf(
             "octo",
             "build-information",
-            "--server=$octopusUrlValue",
-            "--apiKey=$apiKeyValue",
+            "--server=${octopusUrl.get()}",
+            "--apiKey=${apiKey.get()}",
             "--file",
             buildInformation?.absolutePath ?: error("couldn't find build-information.json"),
             "--package-id",
-            packageNameValue,
-            "--version=$versionValue",
+            packageName.get(),
+            "--version=${version.get()}",
             overwriteMode?.let { "--overwrite-mode=$it" }
         ).filterNotNull().joinToString(" ").let { shell(it) }
         if (exitCode == 0) {
@@ -80,15 +59,6 @@ open class UploadBuildInformationTask : DefaultTask() {
             logger.error(inputText)
             throw IllegalStateException("octo build-information exitCode: $exitCode")
         }
-        if (awaitReleases.not()) return
-        awaitReleaseLogic(
-            octopusUrlValue = octopusUrlValue,
-            apiKeyValue = apiKeyValue,
-            waitTimeoutSeconds = waitTimeoutSeconds.getOrElse(600),
-            delayBetweenChecksSeconds = delayBetweenChecksSeconds.getOrElse(5),
-            apiLogLevel = httpLogLevel.getOrElse(HttpLoggingInterceptor.Level.NONE),
-            checkLogic = { determinAnyOngoingTask(packageNameValue = packageNameValue, versionValue = versionValue) }
-        )
     }
